Rewrite of Database layer, split up logics for easier work
