+++ src/lib.rs
@@ -1,11 +1,12 @@
 pub mod common;
 pub mod config;
-pub mod logging;
+pub mod database;
+pub mod databases;
 pub mod handlers;
 pub mod http_api;
 pub mod http_service;
-pub mod udp_common;
-pub mod udp_service;
-pub mod databases;
+pub mod logging;
 pub mod tracker;
 pub mod tracker_objects;
+pub mod udp_common;
+pub mod udp_service;
